+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixes
 - Several fixes for sign up related to lifecycle misusage.
+- Fixed EventManager incorrect isStarted value (now rely on Job.isActive).
 
 ## [7.1.7]